Cleaning the results.php html. Fixing bugs.
I fixed a lot of odd things... The javascript appending still sucks, but
it is a little better and cleaner. It actually puts things in divs... I
fixed the display of the date with javascript. Not sure why that was
broken.

This project still needs lots of cleanup, but I don't think anyone cares
enough to do it.